pharmgeno-ai/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
asyncpg>=0.29.0
pgvector>=0.2.4
redis==5.0.1
httpx==0.26.0
anthropic==0.18.1
python-dotenv==1.0.0
numpy>=1.26.4
pandas>=2.1.4
